Model.scala découpé, à voir si ça marche...
