+++ src/main/scala/model/Model.scala
@@ -1,130 +1,11 @@
 
 package model
 
+import model.town
+import model.road
+import model.point
+import model.train
 import scala.math._
-import scalafx.beans.property.{ DoubleProperty}
-
-
-// A class to represent 2D points
-class Point(var x : Double,var y : Double){
-  def x_coord():Double = {x}
-  def y_coord():Double = {y}
-  override def toString() = s"x : $x ; y : $y"
-  def print():Unit = {println(s"x : $x ; y : $y\n")}
-  def +(that: Point) = {new Point(x + that.x, y + that.y)}  // `this` might be useful
-  def -(that: Point) = {new Point(x - that.x, y - that.y)}
-  def scale(scalar : Double) = {new Point(scalar * x, scalar * y)}
-  def norm() = {this.distance(new Point(0,0))}
-  def distance(p: Point): Double = {
-    sqrt(pow(x-p.x_coord(),2) + pow(y-p.y_coord(),2))
-  }
-  def normalize()= {
-    if (norm() != 0) { this.scale(1/norm()) }
-    this
-  }
-}
-
-
-// A class to represent road between two Towns
-class Road(val begin : Town,val end : Town){
-  val townsVec = end.position - begin.position
-  val length = townsVec.norm()
-  var trainsAB = Seq[Train]()
-  var trainsBA = Seq[Train]()
-  def getStart() = {begin}
-  def getEnd() = {end}
-  def numberOfTrains() = {trainsAB.length + trainsBA.length}
-  def _posTrain(t : Train, b : Boolean) = {
-    var distToBegin = t.distanceOnRoad
-    if (!b) {distToBegin = length - distToBegin}
-    begin.position + townsVec.scale(distToBegin/length)
-  }
-  def getTrainsPos() = {
-    (trainsAB.map(_posTrain(_,true)))++(trainsBA.map(_posTrain(_,false)))
-  }
-  def update() = {
-    var arrived = Seq[(Train,Int)]()
-    for (train <- trainsAB){
-      var dist = train.update()
-      if (dist >= length){
-        arrived = arrived :+ ((train, end.getID()))
-        trainsAB = trainsAB.filter(_ != train)
-      }
-    }
-    for (train <- trainsBA){
-      var dist = train.update()
-      if (dist >= length){
-        arrived = arrived :+ ((train, begin.getID()))
-        trainsBA = trainsBA.filter(_ != train)
-      }
-    }
-  arrived
-  }
-
-  def launchTrain(train : Train, destination : Town) ={
-    if (destination == end){
-      trainsAB = trainsAB :+ train
-      }
-    else{
-      trainsBA = trainsBA :+ train
-      }
-  }
-}
-
-//a class to specify the goods that are available : it returns their name and their price.
-class Goods (val namegoods : String, val pricetag : Double ){}
-
-
-
-// a class to implement the towns of the graphs with information on the name, the population, their wealth and methods to update them when a train come over
-class Town(val id : Int,
-  val name: String,
-  var pop : Int,
-  val listofgoods :List[Goods],
-  // val leaving_roads : List[(Town)],
-  // val coming_roads : List[(Town)],
-  var pos : Point)
-  {
-  override def toString() = {name}
-  var railwayStation = List[Train]()
-  def getID() : Int = {id}
-  def getName() : String = {name}
-  def position() : Point={pos}
-  def population() : Int={pop}
-  def deltaPopulation(delta : Int) = {pop += delta}
-  def incrPop() = {pop = pop+50}
-  def update(){
-  //  pop += 20
-  }
-  def welcomeTrain(train : Train) = {
-    railwayStation = train :: railwayStation
-    pop += train.loading;
-    train.loading = 0
-  }
-  def hasTrains() : Boolean = {railwayStation.isEmpty}
-  def goodbyeTrain(train : Train) : Boolean = {
-    val n = railwayStation.length
-    railwayStation = railwayStation.filter(_!=train)
-    (n != railwayStation.length)
-  }
-}
-
-
-
-//a class to represent the train, for a train, we need to know its speed, its destination, and we need a way to update its information
-class Train(val speed : Double, val name : String){
-  override def toString() = {name}
-  var distanceOnRoad : Double = -1
-  var destination = -1 // L'ID de la destination
-  def update() = { {distanceOnRoad += speed};
-                    distanceOnRoad}
-  def resetDistance() = {distanceOnRoad = 0}
-  def getDestination() = {destination}
-  def setDestination(town : Town) = {destination = town.getID()}
-  def getName() = {name}
-  var loading =  0 // number of passengers in the train.
-  def setLoading( l : Int ) = { loading = l}
-}
 
 
 // Eventually a class to launch a game.
