+++ src/main/scala/model/goods.scala
@@ -1,0 +1,4 @@
+package good
+
+//a class to specify the goods that are available : it returns their name and their price.
+class Good (val namegoods : String, val pricetag : Double ){}
